refact: move db conig info to config.js
