+++ config/config.js
@@ -1,0 +1,20 @@
+const DB_URI = "[Your DB URI]";
+const MOCK_DB_URI = "[Your DB URI]";
+
+const DB_CONFIG = {
+  useCreateIndex: true,
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  useFindAndModify: false
+}
+
+const MOCK_DB_CONFIG = {
+  useCreateIndex: true,
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  useFindAndModify: false
+}
+
+module.exports = { 
+  DB_URI, MOCK_DB_URI, DB_CONFIG, MOCK_DB_CONFIG 
+};
