+++ swagger/db/index.js
@@ -1,12 +1,5 @@
 const mongoose = require('mongoose');
-const DB_URI = "your DB URI";
-
-const DB_CONFIG = {
-  useCreateIndex: true,
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  useFindAndModify: false
-}
+const config = require('../../config/config')
 
 const connect = () => {
   return new Promise((resolve, reject) => {
@@ -15,14 +8,14 @@ const connect = () => {
       const mockgoose = new Mockgoose(mongoose);
       mockgoose.prepareStorage()
         .then(() => {
-          mongoose.connect(DB_URI, DB_CONFIG)
+          mongoose.connect(config.MOCK_DB_URIDB_URI, config.MOCK_DB_CONFIG)
           .then((res, err) => {
             if (err) return reject(err);
             resolve();
           })
         });
     } else {
-      mongoose.connect(DB_URI, DB_CONFIG)
+      mongoose.connect(config.DB_URI, config.DB_CONFIG)
       .then((res, err) => {
         if (err) return reject(err);
         resolve();
